+++ CHANGELOG.md
@@ -30,6 +30,8 @@ The following changes have been completed, and will be included in the next tagg
 * Added `isObjectish()`
 * Added `mustBeObjectish()`
 * Added `validateObjectish()`
+* Added `mustImplementHas()`
+* Added `validateImplementsHas()`
 * Added `off`, `nack` and `narp` as recognised `false` values to the `DEFAULT_BOOLEANISH_RULES`
 * Added `on`, `ack`, `yarp` as recognised `true` values to the `DEFAULT_BOOLEANISH_RULES`
 * Added `Prototypes` to act as module for prototype-related functions
@@ -39,6 +41,7 @@ The following changes have been completed, and will be included in the next tagg
 ### Fixes
 
 * `isEntity()` is now built on `validateEntity()`
+* `implementsHas()` is now built on `validateImplementsHas()`
 * `implementsProtocol()` is now built on `validateImplementsProtocol()`
 
 ### Refactors